% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{arrivals_by_replication}
\alias{arrivals_by_replication}
\title{Number of arrivals in each replication}
\usage{
arrivals_by_replication(envs)
}
\arguments{
\item{envs}{simmer environments}
}
\value{
data.frame
}
\description{
Returns the number of arrivals by replication as a data.frame
}
