% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{create_arrival_generator}
\alias{create_arrival_generator}
\title{Create a trajectory to generate trauma and not-trauma arrivals to the treatment centre.}
\usage{
create_arrival_generator(exp)
}
\arguments{
\item{exp}{An experiment in list form - contains all model parameters. 
Use treat.sim::create_experiment() to generate an experiment list.}
}
\value{
a simmer trajectory
}
\description{
Simulates the arrival of a patient, assigns a patient type and selects 
treatment trajectory.
}
\examples{
default_exp <- create_experiment()
patient_generator <- create_arrival_generator(default_exp)
}
\seealso{
[create_experiment()] to create list containing default and custom experimental parameters.
}
