% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiment.R
\name{create_experiment}
\alias{create_experiment}
\title{Create a simulation experiment parameter list}
\usage{
create_experiment(
  n_triage_bays = DEFAULT_N_TRIAGE,
  n_reg_clerks = DEFAULT_N_REG,
  n_exam_rooms = DEFAULT_N_EXAM,
  n_trauma_rooms = DEFAULT_N_TRAUMA,
  n_non_trauma_cubicles = DEFAULT_NON_TRAUMA_CUBICLES,
  n_trauma_cubicles = DEFAULT_TRAUMA_CUBICLES,
  triage_mean = DEFAULT_TRIAGE_MEAN,
  stabilisation_mean = DEFAULT_TRAUMA_MEAN,
  trauma_treat_params = DEFAULT_TRAUMA_TREATMENT_PARAMS,
  reg_params = DEFAULT_REG_PARAMS,
  exam_params = DEFAULT_EXAM_PARAMS,
  prob_non_trauma_treat = DEFAULT_NON_TRAUMA_TREAT_P,
  nontrauma_treat_params = DEFAULT_NON_TRAUMA_TREATMENT_PARAMS,
  prob_trauma = DEFAULT_PROB_TRAUMA,
  arrival_profile = nelson_arrivals,
  log_level = LOG_LEVEL
)
}
\arguments{
\item{n_triage_bays}{Number of triage bays}

\item{n_reg_clerks}{Number of booking clerks}

\item{n_exam_rooms}{Number of exam rooms}

\item{n_trauma_rooms}{Number of trauma rooms for stabilisation}

\item{n_non_trauma_cubicles}{Number of non-trauma treatment cubicles}

\item{n_trauma_cubicles}{Number of trauma treatment cubicles}

\item{triage_mean}{Mean triage duration (exponential distribution)}

\item{stabilisation_mean}{Mean trauma stabilisation time (exponential distribution)}

\item{trauma_treat_params}{list - mu and sigma for trauma treatment (lognormal)}

\item{reg_params}{list - mu and signma for registration (lognormal)}

\item{exam_params}{list mu and sigma for examination time (normal)}

\item{prob_non_trauma_treat}{probability trauma patient requires treatment}

\item{nontrauma_treat_params}{list - mu and signma for non trauma cubicle treatment (lognormal)}

\item{prob_trauma}{probability arrival has trauma injuries}

\item{arrival_profile}{time dependent arrival profile
The default value used is the package internal datset `nelson_arrivals`. 
A new dataset should include three columns: 
1. period (equally spaced 60 min intervals 0 to 1020)
2. arrival_rate (per hour)
3. arrival_rate2 (per minute)}

\item{log_level}{simmer log level. Set to 0 to hide all debug info as model runs.}
}
\value{
A list
}
\description{
Create and return a list containing all of the parameters used by the 
treat.sim model.
}
\details{
If no parameters are passed to the function then a default experiment
is created.

Users can choose to create custom experiments by passing values to 
the corresponding parameters in the function.
}
\examples{
# sample from Nelson arrivals at time 20.0
default_experiment <- create_experiment()

# set number of triage bays to 3
default_experiment <- create_experiment(n_triage_bays=3)
}
