% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{create_non_trauma_pathway}
\alias{create_non_trauma_pathway}
\title{Create and return a trajectory for non-trauma patients}
\usage{
create_non_trauma_pathway(exp)
}
\arguments{
\item{exp}{An experiment in list form - contains all model parameters. 
Use treat.sim::create_experiment() to generate an experiment list.}
}
\value{
a simmer trajectory
}
\description{
Simulates the process for a non-trauma patient. Logic is:

1. Triage (requires triage bay)

2. Registriation (requires clert)

3. Examination (requires exam room)

4. Probabilistic decision about treatment
 
4.1. Treatment (requires cubicle)

5. Discharge
}
\examples{
default_exp <- create_experiment()
nt_traj <- create_non_trauma_pathway(default_exp)
}
\seealso{
[create_experiment()] to create list containing default and custom experimental parameters.
}
