% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{create_nt_cubicle_treatment}
\alias{create_nt_cubicle_treatment}
\title{Create and return a simmer::trajectory representing cubicle treatment of non-trauma patients.}
\usage{
create_nt_cubicle_treatment(exp)
}
\arguments{
\item{exp}{An experiment in list form - contains all model parameters. 
Use treat.sim::create_experiment() to generate an experiment list.}
}
\value{
a simmer trajectory
}
\description{
Simulates cubicle treatment of trauma patients (log normally distributed time)
}
\examples{
default_exp <- create_experiment()
nt_treatement_traj <- create_nt_cubicle_treatment(default_exp)
}
\seealso{
[create_experiment()] to create list containing default and custom experimental parameters.
}
