% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{create_summary_table}
\alias{create_summary_table}
\title{Create a summary table of multiple replications data}
\usage{
create_summary_table(rep_table, dp = 2)
}
\arguments{
\item{rep_table}{data.frame containing replications (rows) and KPIs (cols)}

\item{dp}{the number of decimal places}
}
\value{
data.frame
}
\description{
Accepts a table of replication results and returns the mean of those
values in a data.frame
}
\seealso{
[replication_results_table()] to create a table of replications
}
