% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{create_trauma_pathway}
\alias{create_trauma_pathway}
\title{Create and return a trauma pathway simmer trajectory}
\usage{
create_trauma_pathway(exp)
}
\arguments{
\item{exp}{An experiment in list form - contains all model parameters. 
Use treat.sim::create_experiment() to generate an experiment list.}
}
\value{
a simmer trajectory
}
\description{
Trauma patients follow this process in the trajectory:

1. Triage (requires triage bay)

2. Stabilisation (requires trauma room)

3. Treatment (requirement trauma cubicle)

4. Discharge
}
\examples{
default_exp <- create_experiment()
trauma_traj <- create_trauma_pathway(default_exp)
}
\seealso{
[create_experiment()] to create list containing default and custom experimental parameters.
}
