% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_of_replications.R
\name{histogram_of_replications}
\alias{histogram_of_replications}
\title{Histogram of replications for a selected KPI}
\usage{
histogram_of_replications(rep_table, column_name, unit_label, n_bins = 10)
}
\arguments{
\item{rep_table}{data.frame containing replications (rows) and KPIs (cols)}

\item{column_name}{string name of the KPI to plot}

\item{unit_label}{string of the x-axis label unit}

\item{n_bins}{number of bins for the histogram}
}
\value{
plot
}
\description{
Accepts a table of replication results and a ggplot histogram object
for a selected column.
}
\seealso{
[replication_results_table()] to create a table of replications
}
