% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nelson_arrivals}
\alias{nelson_arrivals}
\title{Time dependent arrival profile from Nelson (2013)}
\format{
## `nelson_arrivals`
This is a data.frame containing the following columns
\describe{
    \item{period}{0 to 480 in 60 minute intervals}
    \item{arrival_rate}{rate per hour}
    \item{arrival_rate2}{rate per minute}
}
}
\source{
: https://blogs.sas.com/content/iml/2014/06/04/simulate-lognormal-data-with-specified-mean-and-variance.html
}
\usage{
nelson_arrivals(path = NSPP_PATH)
}
\arguments{
\item{path}{A string. Optional path to arrival Profile from Nelson (2013)}
}
\value{
A data.frame
}
\description{
A simple arrival profile for use with treat.sim
}
