% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nspp_thinning.R
\name{nspp_thinning}
\alias{nspp_thinning}
\title{Non-stationary Poisson Process (NSPP) via thinning}
\usage{
nspp_thinning(simulation_time, arrival_profile, debug = FALSE)
}
\arguments{
\item{simulation_time}{A number. The current simulation time}

\item{arrival_profile}{A data.frame The time dependent arrival profile}

\item{debug}{bool. 
TRUE = printout log of thinning after a sample has been accepted.
FALSE = no debug info provided.}
}
\value{
A number
}
\description{
Use the current simulation time to sample an appropriate
inter-arrival time from a user set NSPP.  Assumes equal spaced intervals.
}
\details{
Thinning is an acceptance-rejection approach to sampling
inter-arrival times (IAT) from a time dependent distribution
where each time period follows its own exponential distribution.

There are two random variables employed in sampling: an exponential
distribution (used to sample IAT) and a uniform distribution (used to accept/reject samples).

All IATs are sampled from an Exponential distribution with the highest
arrival rate (most frequent). These arrivals are then rejected (thinned)
proportional to the ratio of the current arrival rate to the maximum
arrival rate.  The algorithm executes until a sample is accepted. The IAT
returned is the sum of all the IATs that were sampled.
}
\examples{
# sample from Nelson arrivals at time 20.0
nspp_thinning(20.0, nelson_arrivals, debug=TRUE)
}
