% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{replication_results_table}
\alias{replication_results_table}
\title{Function to create the replications table}
\usage{
replication_results_table(
  reps,
  exp,
  results_collection_period = DEFAULT_RESULTS_COLLECTION_PERIOD
)
}
\arguments{
\item{reps}{list of simmer environments}

\item{exp}{list of "experiment" contains all of the model 
parameters used to create the results}

\item{results_collection_period}{the length of time results were collected.}
}
\value{
data.frame
}
\description{
Accepts a list of simmer environments and converts to a data,frame
of replications (rows) x KPIs (cols).
}
