% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{resource_utilisation_by_replication}
\alias{resource_utilisation_by_replication}
\title{Resource utilisation by replication}
\usage{
resource_utilisation_by_replication(reps, exp, results_collection_period)
}
\arguments{
\item{exp}{list of parameters for experiment}

\item{results_collection_period}{results collection simulation time}

\item{envs}{simmer environments}
}
\value{
data.frame
}
\description{
Returns the  utilisation of resource by replication as a data.frame
calculation:
total busy time / total scheduled resource time.
where total scheduled time = n_resource * results collection period.
}
