% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{resource_waiting_times_by_replication}
\alias{resource_waiting_times_by_replication}
\title{Waiting times for resources by replication}
\usage{
resource_waiting_times_by_replication(reps)
}
\arguments{
\item{envs}{simmer environments from each replication of the model}
}
\value{
data.frame
}
\description{
Returns the mean waiting times for resourcse by replication as a data.frame
}
