% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_arrival_type}
\alias{sample_arrival_type}
\title{Sample a patient type}
\usage{
sample_arrival_type(p, n = 1)
}
\arguments{
\item{p}{A number: the probability a patient has trauma on arrival}

\item{n}{A number: number of samples to generate.}
}
\description{
`sample_arrival_type` samples if a patient type is trauma or non-trauma
with a given probability.
}
\details{
The function uses the Bernoulli distribution (Rlab) to sample
if a patient is Trauma or Non-Trauma.  The return values are 
1 = Trauma, 2 = Non-trauma.
}
\examples{
patient_type = sample_arrival_type(0.4)

}
