% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_nt_trauma_treatment}
\alias{sample_nt_trauma_treatment}
\title{Sample a if a non-trauma patient requires treatment}
\usage{
sample_nt_trauma_treatment(p)
}
\arguments{
\item{p}{A number: The probability the patient requires treatment}
}
\description{
`sample_nt_trauma_treatment` samples if a non-trauma patient
requires cubicle treatment
}
\details{
The function uses the Bernouli distribution (Rlab) to sample
if a patient is requires treatment or not.  The return values are 
1 = Treatment, 0 = No treatment
}
\examples{
treat_flag = sample_nt_trauma_treatment(0.30)

}
