% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{single_run}
\alias{single_run}
\title{Perform a single replication of the model}
\usage{
single_run(
  env,
  exp,
  rep_number = 1,
  run_length = DEFAULT_RESULTS_COLLECTION_PERIOD,
  debug_arrivals = FALSE
)
}
\arguments{
\item{env}{simmer:environment}

\item{exp}{An experiment in list form - contains all model parameters.}

\item{rep_number}{the replication number (default=1)}

\item{run_length}{run length of the simulation}

\item{debug_arrivals}{boolean flag to show debug info for the thinning process
Use treat.sim::create_experiment() to generate an experiment list.}
}
\value{
a simmer environment
}
\description{
The function adds treat.sim resources to an environment.
The model is then run for a single replication. The environment is
returned for use in results analysis.
}
\examples{
set.seed(42)
exp <- create_experiment(log_level=0)
treat_sim <- simmer::simmer("TreatSim", log_level=exp$log_level)
treat_sim <- single_run(treat_sim, exp)
print("Simulation Complete.")
}
\seealso{
[create_experiment()] to create list containing default and custom experimental parameters.
}
