% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{system_kpi_for_rep_i}
\alias{system_kpi_for_rep_i}
\title{System level KPIs}
\usage{
system_kpi_for_rep_i(reps, rep_i)
}
\arguments{
\item{reps}{list of simmer environments}

\item{rep_i}{replication for calculation}
}
\value{
data.frame
}
\description{
Calculates mean time in system and throughput for a replication
}
