% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_moments_from_lognormal.R
\name{normal_moments_from_lognormal}
\alias{normal_moments_from_lognormal}
\title{Mean and Variance of the underlying Normal Distribution}
\usage{
normal_moments_from_lognormal(mean, std)
}
\arguments{
\item{mean}{A number. Sample mean.}

\item{stdev}{A number. Sample standard deviation}
}
\value{
A list
}
\description{
`normal_moments_from_lognormal` calculates the mu and sigma
of the normal distribution underlying a lognormal
mean and standard
}
\details{
`rlnorm` from `stats` is designed to sample from the lognormal distribution. 
The parameters is expects are moments of the underlying normal distribution
Using sample mean and standard deviation this function calculates 
the mu and sigma of the normal distribution. 
source: https://blogs.sas.com/content/iml/2014/06/04/simulate-lognormal-data-with-specified-mean-and-variance.html
}
\examples{
normal_moments_from_lognormal(mean = 125.0,
                              std = 5.0)
}
